refactor: split api/queries.py into a per-domain package (Stage 2)
The 2217-line god module becomes src/mondo/api/queries/ with one module
per domain (items, boards, docs, columns, groups, workspaces, users,
teams, subitems, updates, webhooks, files, activity, folders, favorites,
tags, me). __init__.py re-exports the full public surface so every
'from mondo.api.queries import X' keeps working unchanged. Zero behaviour
change (public-API parity verified). Also drops the now-false 'query set
is small and lives in one place' comment.
