+++ src/mondo/api/queries/__init__.py
@@ -1,0 +1,271 @@
+"""GraphQL queries and mutations used by mondo commands.
+
+Kept inline as string constants (not `.graphql` files). Split into one
+module per domain for maintainability; this package re-exports the full
+public surface so `from mondo.api.queries import X` keeps working unchanged.
+All mutations use variables to avoid the double-JSON escape trap
+(monday-api.md §11.4).
+"""
+
+from __future__ import annotations
+
+from .activity import BOARD_ACTIVITY_LOGS
+from .boards import (
+    BOARD_ARCHIVE,
+    BOARD_CREATE,
+    BOARD_DELETE,
+    BOARD_DUPLICATE,
+    BOARD_GET,
+    BOARD_GET_WITH_VIEWS,
+    BOARD_ITEMS_COUNT,
+    BOARD_SET_PERMISSION,
+    BOARD_UPDATE,
+    BOARD_UPDATE_HIERARCHY,
+    build_boards_list_query,
+)
+from .columns import (
+    CHANGE_COLUMN_VALUE,
+    CHANGE_MULTIPLE_COLUMN_VALUES,
+    COLUMN_CHANGE_METADATA,
+    COLUMN_CONTEXT,
+    COLUMN_CREATE,
+    COLUMN_DELETE,
+    COLUMN_RENAME,
+    COLUMNS_ON_BOARD,
+    CREATE_OR_GET_TAG,
+)
+from .docs import (
+    ADD_CONTENT_TO_DOC_FROM_MARKDOWN,
+    CREATE_DOC_BLOCK,
+    CREATE_DOC_IN_WORKSPACE,
+    CREATE_DOC_ON_ITEM,
+    DELETE_DOC,
+    DELETE_DOC_BLOCK,
+    DOC_GET_BY_ID,
+    DOC_GET_BY_ID_BLOCKS_PAGE,
+    DOC_HEAD_BY_OBJECT_ID,
+    DOC_VERSION_DIFF,
+    DOC_VERSION_HISTORY,
+    DOCS_BY_OBJECT_ID,
+    DOCS_BY_OBJECT_ID_BLOCKS_PAGE,
+    DUPLICATE_DOC,
+    EXPORT_MARKDOWN_FROM_DOC,
+    IMPORT_DOC_FROM_HTML,
+    UPDATE_DOC_BLOCK,
+    UPDATE_DOC_NAME,
+    build_docs_list_query,
+)
+from .favorites import FAVORITES_LIST
+from .files import ASSETS_GET, FILE_UPLOAD_ITEM, FILE_UPLOAD_UPDATE
+from .folders import (
+    FOLDER_CREATE,
+    FOLDER_DELETE,
+    FOLDER_GET,
+    FOLDER_UPDATE,
+    build_folders_list_query,
+)
+from .groups import (
+    GROUP_ARCHIVE,
+    GROUP_CREATE,
+    GROUP_DELETE,
+    GROUP_DUPLICATE,
+    GROUP_UPDATE,
+    GROUPS_LIST,
+)
+from .items import (
+    ITEM_ARCHIVE,
+    ITEM_BOARD_LOOKUP,
+    ITEM_CREATE,
+    ITEM_DELETE,
+    ITEM_DUPLICATE,
+    ITEM_GET,
+    ITEM_GET_WITH_COLUMNS,
+    ITEM_GET_WITH_SUBITEMS,
+    ITEM_GET_WITH_UPDATES,
+    ITEM_MOVE_BOARD,
+    ITEM_MOVE_GROUP,
+    ITEM_RENAME,
+    ITEMS_PAGE_INITIAL,
+    ITEMS_PAGE_INITIAL_WITH_SUBITEMS,
+    ITEMS_PAGE_NEXT,
+    ITEMS_PAGE_NEXT_WITH_SUBITEMS,
+    build_items_page_queries,
+)
+from .me import (
+    ACCOUNT_ONLY,
+    AGGREGATE_BOARD,
+    CREATE_NOTIFICATION,
+    ME_FULL,
+    ME_QUERY,
+    VALIDATIONS_LIST,
+)
+from .subitems import SUBITEM_CREATE, SUBITEM_GET, SUBITEMS_LIST
+from .tags import TAG_BY_BOARD, TAGS_LIST
+from .teams import (
+    ASSIGN_TEAM_OWNERS,
+    REMOVE_TEAM_OWNERS,
+    TEAM_CREATE,
+    TEAM_DELETE,
+    TEAMS_LIST,
+)
+from .updates import (
+    UPDATE_CLEAR_ITEM,
+    UPDATE_CREATE,
+    UPDATE_DELETE,
+    UPDATE_EDIT,
+    UPDATE_GET,
+    UPDATE_LIKE,
+    UPDATE_PIN,
+    UPDATE_UNLIKE,
+    UPDATE_UNPIN,
+    UPDATES_FOR_ITEM,
+    UPDATES_LIST_PAGE,
+)
+from .users import (
+    ADD_USERS_TO_TEAM,
+    REMOVE_USERS_FROM_TEAM,
+    USER_GET,
+    USERS_ACTIVATE,
+    USERS_DEACTIVATE,
+    USERS_LIST_PAGE,
+    USERS_UPDATE_AS_ADMINS,
+    USERS_UPDATE_AS_GUESTS,
+    USERS_UPDATE_AS_MEMBERS,
+    USERS_UPDATE_AS_VIEWERS,
+)
+from .webhooks import WEBHOOK_CREATE, WEBHOOK_DELETE, WEBHOOKS_LIST
+from .workspaces import (
+    WORKSPACE_ADD_TEAMS,
+    WORKSPACE_ADD_USERS,
+    WORKSPACE_CREATE,
+    WORKSPACE_DELETE,
+    WORKSPACE_GET,
+    WORKSPACE_REMOVE_TEAMS,
+    WORKSPACE_REMOVE_USERS,
+    WORKSPACE_UPDATE,
+    WORKSPACES_LIST_PAGE,
+)
+
+__all__ = [
+    "ACCOUNT_ONLY",
+    "ADD_CONTENT_TO_DOC_FROM_MARKDOWN",
+    "ADD_USERS_TO_TEAM",
+    "AGGREGATE_BOARD",
+    "ASSETS_GET",
+    "ASSIGN_TEAM_OWNERS",
+    "BOARD_ACTIVITY_LOGS",
+    "BOARD_ARCHIVE",
+    "BOARD_CREATE",
+    "BOARD_DELETE",
+    "BOARD_DUPLICATE",
+    "BOARD_GET",
+    "BOARD_GET_WITH_VIEWS",
+    "BOARD_ITEMS_COUNT",
+    "BOARD_SET_PERMISSION",
+    "BOARD_UPDATE",
+    "BOARD_UPDATE_HIERARCHY",
+    "CHANGE_COLUMN_VALUE",
+    "CHANGE_MULTIPLE_COLUMN_VALUES",
+    "COLUMNS_ON_BOARD",
+    "COLUMN_CHANGE_METADATA",
+    "COLUMN_CONTEXT",
+    "COLUMN_CREATE",
+    "COLUMN_DELETE",
+    "COLUMN_RENAME",
+    "CREATE_DOC_BLOCK",
+    "CREATE_DOC_IN_WORKSPACE",
+    "CREATE_DOC_ON_ITEM",
+    "CREATE_NOTIFICATION",
+    "CREATE_OR_GET_TAG",
+    "DELETE_DOC",
+    "DELETE_DOC_BLOCK",
+    "DOCS_BY_OBJECT_ID",
+    "DOCS_BY_OBJECT_ID_BLOCKS_PAGE",
+    "DOC_GET_BY_ID",
+    "DOC_GET_BY_ID_BLOCKS_PAGE",
+    "DOC_HEAD_BY_OBJECT_ID",
+    "DOC_VERSION_DIFF",
+    "DOC_VERSION_HISTORY",
+    "DUPLICATE_DOC",
+    "EXPORT_MARKDOWN_FROM_DOC",
+    "FAVORITES_LIST",
+    "FILE_UPLOAD_ITEM",
+    "FILE_UPLOAD_UPDATE",
+    "FOLDER_CREATE",
+    "FOLDER_DELETE",
+    "FOLDER_GET",
+    "FOLDER_UPDATE",
+    "GROUPS_LIST",
+    "GROUP_ARCHIVE",
+    "GROUP_CREATE",
+    "GROUP_DELETE",
+    "GROUP_DUPLICATE",
+    "GROUP_UPDATE",
+    "IMPORT_DOC_FROM_HTML",
+    "ITEMS_PAGE_INITIAL",
+    "ITEMS_PAGE_INITIAL_WITH_SUBITEMS",
+    "ITEMS_PAGE_NEXT",
+    "ITEMS_PAGE_NEXT_WITH_SUBITEMS",
+    "ITEM_ARCHIVE",
+    "ITEM_BOARD_LOOKUP",
+    "ITEM_CREATE",
+    "ITEM_DELETE",
+    "ITEM_DUPLICATE",
+    "ITEM_GET",
+    "ITEM_GET_WITH_COLUMNS",
+    "ITEM_GET_WITH_SUBITEMS",
+    "ITEM_GET_WITH_UPDATES",
+    "ITEM_MOVE_BOARD",
+    "ITEM_MOVE_GROUP",
+    "ITEM_RENAME",
+    "ME_FULL",
+    "ME_QUERY",
+    "REMOVE_TEAM_OWNERS",
+    "REMOVE_USERS_FROM_TEAM",
+    "SUBITEMS_LIST",
+    "SUBITEM_CREATE",
+    "SUBITEM_GET",
+    "TAGS_LIST",
+    "TAG_BY_BOARD",
+    "TEAMS_LIST",
+    "TEAM_CREATE",
+    "TEAM_DELETE",
+    "UPDATES_FOR_ITEM",
+    "UPDATES_LIST_PAGE",
+    "UPDATE_CLEAR_ITEM",
+    "UPDATE_CREATE",
+    "UPDATE_DELETE",
+    "UPDATE_DOC_BLOCK",
+    "UPDATE_DOC_NAME",
+    "UPDATE_EDIT",
+    "UPDATE_GET",
+    "UPDATE_LIKE",
+    "UPDATE_PIN",
+    "UPDATE_UNLIKE",
+    "UPDATE_UNPIN",
+    "USERS_ACTIVATE",
+    "USERS_DEACTIVATE",
+    "USERS_LIST_PAGE",
+    "USERS_UPDATE_AS_ADMINS",
+    "USERS_UPDATE_AS_GUESTS",
+    "USERS_UPDATE_AS_MEMBERS",
+    "USERS_UPDATE_AS_VIEWERS",
+    "USER_GET",
+    "VALIDATIONS_LIST",
+    "WEBHOOKS_LIST",
+    "WEBHOOK_CREATE",
+    "WEBHOOK_DELETE",
+    "WORKSPACES_LIST_PAGE",
+    "WORKSPACE_ADD_TEAMS",
+    "WORKSPACE_ADD_USERS",
+    "WORKSPACE_CREATE",
+    "WORKSPACE_DELETE",
+    "WORKSPACE_GET",
+    "WORKSPACE_REMOVE_TEAMS",
+    "WORKSPACE_REMOVE_USERS",
+    "WORKSPACE_UPDATE",
+    "build_boards_list_query",
+    "build_docs_list_query",
+    "build_folders_list_query",
+    "build_items_page_queries",
+]
